+++ CHANGES.md
@@ -1,6 +1,7 @@
 Version 0.2.3 released 2012-6-26
 ----------------------------------------------------------------------
 
+* Fixed query support for URL class (issue #61)
 * Improved logging. No root logger hijacking anymore (issue #62)
 * Fixed job.Description.number_of_processes (issue #63)